Fixed issue#001
Replaced search function with findall function to return an list containing all matches which are concatenated into result. If result is not none it is returned and if it is a blank object is returned.  Further, the write to file is edited to include a test if incoming data is none or not and based on that writing is done .